requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
loguru==0.7.2
instaloader==4.11
sqlalchemy==2.0.23
pytest==7.4.4
